Implements transaction view model mappers.
